DROOLS-4343 Change fact handle id type to long (#2468)
* DROOLS-4343 Change fact handle id type to long

* DROOLS-4343 Change fact handle id type to long

+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/marshalling/util/OldOutputMarshallerMethods.java
@@ -72,7 +72,7 @@ public class OldOutputMarshallerMethods {
             int type,
             InternalFactHandle handle) throws IOException {
         stream.writeInt( type );
-        stream.writeInt( handle.getId() );
+        stream.writeLong( handle.getId() );
         stream.writeLong( handle.getRecency() );
 
         if ( type == 2) {
